# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: longue-vue CMDB API
  version: 0.1.0
  summary: Kubernetes-centric CMDB aligned with the ANSSI SecNumCloud framework.
  description: |
    REST API for **longue-vue**, a CMDB for Kubernetes environments aligned with the
    ANSSI SecNumCloud (SNC) qualification framework.

    This version (0.1.0) is a walking skeleton covering:
      - Liveness and readiness probes
      - `Cluster` resource CRUD

    Subsequent versions will add Node, Namespace, Workload, Pod, and other
    Kubernetes entity kinds mapped onto the ANSSI cartography layers.

    Errors follow RFC 7807 (`application/problem+json`).

    Interactive API documentation is available at `/docs/` for authenticated
    operators; sign in via `/ui/login` first to use "Try it out".
servers:
  - url: /
    description: Current origin (recommended for browser-based "Try it out" via /docs/)
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/clusters/{id}/flow-references:
    get:
      operationId: listFlowReferences
      summary: List a cluster's reference flow rows
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/FlowReference" } }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createFlowReference
      summary: Create a reference flow row
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowReferenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowReference"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
...
  /v1/flow-references/{id}:
    patch:
      operationId: updateFlowReference
      summary: Update a reference flow row
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlowReferenceInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowReference"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteFlowReference
      summary: Delete a reference flow row
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "204":
          description: Deleted
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: |
        API token issued out-of-band. Send as `Authorization: Bearer <token>`.

        Scopes granted to a token:
          - `read`   — list and get operations
          - `write`  — create and update (implies `read` at the caller's discretion)
          - `delete` — remove resources
          - `admin`  — all of the above; reserved for future admin-only endpoints

        Tokens carrying the `admin` scope implicitly satisfy any scope required
        by an operation.

        Per ADR-0007, machine tokens are created by an admin in the UI
        (see `POST /v1/admin/tokens`) rather than injected through env vars.

    SessionCookie:
      type: apiKey
      in: cookie
      name: longue_vue_session
      description: |
        Human session cookie set by `POST /v1/auth/login`. HttpOnly,
        SameSite=Strict, 8-hour sliding expiry. Per ADR-0007.

        Every operation accepts either `BearerAuth` or `SessionCookie` —
        scope enforcement reads from the caller's role (for sessions) or
        token scopes (for bearer), so the two paths share the same
        downstream authorisation check.
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Token is valid but lacks the scope required by the operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflict with current state (e.g., duplicate name).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      description: RFC 7807 problem details.
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    FlowReference:
      type: object
      description: One operator-declared row of a cluster's reference flow matrix (ADR-0036).
      required: [id, cluster_id, layer, direction, src_kind, src_ref, dst_kind, dst_ref, protocol, justification, created_at, updated_at]
      properties:
        id:            { type: string, format: uuid }
        cluster_id:    { type: string, format: uuid }
        layer:         { type: string, enum: [perimeter, internal] }
        direction:     { type: string, enum: [ingress, egress] }
        src_kind:      { type: string, enum: [workload, service, namespace, endpoint_group, cidr] }
        src_ref:       { type: string }
        dst_kind:      { type: string, enum: [workload, service, namespace, endpoint_group, cidr] }
        dst_ref:       { type: string }
        protocol:      { type: string, enum: [tcp, udp, icmp, any] }
        from_port:     { type: integer, nullable: true }
        to_port:       { type: integer, nullable: true }
        justification: { type: string }
        created_by:    { type: string, format: uuid, nullable: true }
        created_at:    { type: string, format: date-time }
        updated_at:    { type: string, format: date-time }
    FlowReferenceInput:
      type: object
      description: >-
        Create/update payload for a FlowReference. A perimeter flow needs
        exactly one endpoint_group side; an internal flow must reference none.
        Justification is required.
      required: [layer, direction, src_kind, src_ref, dst_kind, dst_ref, protocol, justification]
      properties:
        layer:         { type: string, enum: [perimeter, internal] }
        direction:     { type: string, enum: [ingress, egress] }
        src_kind:      { type: string, enum: [workload, service, namespace, endpoint_group, cidr] }
        src_ref:       { type: string }
        dst_kind:      { type: string, enum: [workload, service, namespace, endpoint_group, cidr] }
        dst_ref:       { type: string }
        protocol:      { type: string, enum: [tcp, udp, icmp, any] }
        from_port:     { type: integer, nullable: true }
        to_port:       { type: integer, nullable: true }
        justification: { type: string }
...